:wrench: CustomerController created and implemented without the service

+++ src/main/kotlin/dev/brunofelix/bookapi/domain/entity/Customer.kt
@@ -6,7 +6,7 @@ import jakarta.persistence.*
 data class Customer(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null,
+    var id: Long? = null,
 
     @Column
     var name: String,
